spilt ai summary component
